[160_common_interface], issue #160, defined tph_common_interface.py for common interfaces of thermophysical properties
