+++ lbh15/properties/lead_properties.py
@@ -4,13 +4,23 @@ from typing import List
 from typing import Union
 import numpy as np
 from scipy.constants import atm
-from .interface import PropertyInterface
+from .tph_common_interface import SaturationVapourPressureInterface
+from .tph_common_interface import SurfaceTensionInterface
+from .tph_common_interface import DensityInterface
+from .tph_common_interface import ThermalExpansionInterface
+from .tph_common_interface import SpeedOfSoundInterface
+from .tph_common_interface import IsentropicCompressibilityInterface
+from .tph_common_interface import SpecificHeatInterface
+from .tph_common_interface import SpecificEnthalpyInterface
+from .tph_common_interface import DynamicViscosityInterface
+from .tph_common_interface import ElectricalResistivityInterface
+from .tph_common_interface import ThermalConductivityInterface
 from .._decorators import range_warning
 from .._commons import LEAD_MELTING_TEMPERATURE as T_m0
 from .._commons import LEAD_BOILING_TEMPERATURE as T_b0
 
 
-class p_s(PropertyInterface):
+class p_s(SaturationVapourPressureInterface):
     """
     Liquid lead *saturation vapour pressure* property class.
     """
@@ -78,20 +88,6 @@ class p_s(PropertyInterface):
         return [T_m0, T_b0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Saturation vapour pressure unit
-        """
-        return "[Pa]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Saturation vapour pressure long name
-        """
-        return "saturation vapour pressure"
-
-    @property
     def description(self) -> str:
         """
         str : Saturation vapour pressure description
@@ -99,7 +95,7 @@ class p_s(PropertyInterface):
         return f"Liquid lead {self.long_name}"
 
 
-class sigma(PropertyInterface):
+class sigma(SurfaceTensionInterface):
     """
     Liquid lead *surface tension* property class.
     """
@@ -144,20 +140,6 @@ class sigma(PropertyInterface):
         return [T_m0, 1300.0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Surface tension unit
-        """
-        return "[N/m]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Surface tension long name
-        """
-        return "surface tension"
-
-    @property
     def description(self) -> str:
         """
         str : Surface tension description
@@ -165,7 +147,7 @@ class sigma(PropertyInterface):
         return f"Liquid lead {self.long_name}"
 
 
-class rho(PropertyInterface):
+class rho(DensityInterface):
     """
     Liquid lead *density* property class.
     """
@@ -217,20 +199,6 @@ class rho(PropertyInterface):
         return [T_m0, T_b0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Density unit
-        """
-        return "[kg/m^3]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Density long name
-        """
-        return "density"
-
-    @property
     def description(self) -> str:
         """
         str : Density description
@@ -238,7 +206,7 @@ class rho(PropertyInterface):
         return f"Liquid lead {self.long_name}"
 
 
-class alpha(PropertyInterface):
+class alpha(ThermalExpansionInterface):
     """
     Liquid lead *thermal expansion coefficient* property class.
     """
@@ -276,20 +244,6 @@ class alpha(PropertyInterface):
         return [T_m0, T_b0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Thermal expansion coefficient unit
-        """
-        return "[1/K]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Thermal expansion coefficient long name
-        """
-        return "thermal expansion coefficient"
-
-    @property
     def description(self) -> str:
         """
         str : Thermal expansion coefficient description
@@ -297,7 +251,7 @@ class alpha(PropertyInterface):
         return f"Liquid lead {self.long_name}"
 
 
-class u_s(PropertyInterface):
+class u_s(SpeedOfSoundInterface):
     """
     Liquid lead *sound velocity* property class.
     """
@@ -342,20 +296,6 @@ class u_s(PropertyInterface):
         return [T_m0, 2000.0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Sound velocity unit
-        """
-        return "[m/s]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Sound velocity long name
-        """
-        return "sound velocity"
-
-    @property
     def description(self) -> str:
         """
         str : Sound velocity description
@@ -363,7 +303,7 @@ class u_s(PropertyInterface):
         return "Sound velocity in liquid lead"
 
 
-class beta_s(PropertyInterface):
+class beta_s(IsentropicCompressibilityInterface):
     """
     Liquid lead *isentropic compressibility* property class.
     """
@@ -402,20 +342,6 @@ class beta_s(PropertyInterface):
         return [T_m0, 2000.0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Isentropic compressibility unit
-        """
-        return "[1/Pa]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Isentropic compressibility long name
-        """
-        return "isentropic compressibility"
-
-    @property
     def description(self) -> str:
         """
         str : Isentropic compressibility description
@@ -423,7 +349,19 @@ class beta_s(PropertyInterface):
         return f"Liquid lead {self.long_name}"
 
 
-class cp_sobolev2011(PropertyInterface):
+class LeadCpInterface(SpecificHeatInterface):
+    """
+    Liquid lead *specific heat capacity* property interface.
+    """
+    @property
+    def name(self) -> str:
+        """
+        str : Name of the property
+        """
+        return "cp"
+
+
+class cp_sobolev2011(LeadCpInterface):
     """
     Liquid lead *specific heat capacity* property class exploiting
     the **sobolev2011** correlation.
@@ -454,26 +392,11 @@ class cp_sobolev2011(PropertyInterface):
         return 176.2 - T * (4.923e-2 - 1.544e-5 * T) - 1.524e6 / T / T
 
     @property
-    def name(self) -> str:
-        """
-        str : Name of the property
-        """
-        return "cp"
-
-    @property
     def correlation_name(self) -> str:
         """
         str : Name of the correlation
         """
         return 'sobolev2011'
-
-    @property
-    def is_injective(self) -> bool:
-        """
-        bool : `True` if the correlation is injective,
-        `False` otherwise.
-        """
-        return False
 
     @property
     def range(self) -> List[float]:
@@ -484,20 +407,6 @@ class cp_sobolev2011(PropertyInterface):
         return [T_m0, 2000.0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Specific heat capacity unit
-        """
-        return "[J/(kg*K)]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Specific heat capacity long name
-        """
-        return "specific heat capacity"
-
-    @property
     def description(self) -> str:
         """
         str : Specific heat capacity description
@@ -505,7 +414,7 @@ class cp_sobolev2011(PropertyInterface):
         return f"Liquid lead {self.long_name}"
 
 
-class cp_gurvich1991(PropertyInterface):
+class cp_gurvich1991(LeadCpInterface):
     """
     Liquid lead *specific heat capacity* property class exploiting
     the **gurvich1991** correlation.
@@ -537,26 +446,11 @@ class cp_gurvich1991(PropertyInterface):
             - 1.524e6 / T / T
 
     @property
-    def name(self) -> str:
-        """
-        str : Name of the property
-        """
-        return "cp"
-
-    @property
     def correlation_name(self) -> str:
         """
         str : Name of the correlation
         """
         return 'gurvich1991'
-
-    @property
-    def is_injective(self) -> bool:
-        """
-        bool : `True` if the correlation is injective,
-        `False` otherwise.
-        """
-        return False
 
     @property
     def range(self) -> List[float]:
@@ -567,20 +461,6 @@ class cp_gurvich1991(PropertyInterface):
         return [T_m0, 2000.0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Specific heat capacity unit
-        """
-        return "[J/(kg*K)]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Specific heat capacity long name
-        """
-        return "specific heat capacity"
-
-    @property
     def description(self) -> str:
         """
         str : Specific heat capacity description
@@ -588,7 +468,7 @@ class cp_gurvich1991(PropertyInterface):
         return f"Liquid lead {self.long_name}"
 
 
-class h(PropertyInterface):
+class h(SpecificEnthalpyInterface):
     """
     Liquid lead *specific enthalpy* property class.
     """
@@ -635,20 +515,6 @@ class h(PropertyInterface):
         return [T_m0, 2000.0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Specific enthalpy unit
-        """
-        return "[J/kg]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Specific enthalpy long name
-        """
-        return "specific enthalpy"
-
-    @property
     def description(self) -> str:
         """
         str : Specific enthalpy description
@@ -658,7 +524,7 @@ class h(PropertyInterface):
                 "the melting point enthalpy)")
 
 
-class mu(PropertyInterface):
+class mu(DynamicViscosityInterface):
     """
     Liquid lead *dynamic viscosity* property class.
     """
@@ -717,20 +583,6 @@ class mu(PropertyInterface):
         return [T_m0, 1473.0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Dynamic viscosity unit
-        """
-        return "[Pa*s]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Dynamic viscosity long name
-        """
-        return "dynamic viscosity"
-
-    @property
     def description(self) -> str:
         """
         str : Dynamic viscosity description
@@ -738,7 +590,7 @@ class mu(PropertyInterface):
         return f"Liquid lead {self.long_name}"
 
 
-class r(PropertyInterface):
+class r(ElectricalResistivityInterface):
     """
     Liquid lead *electrical resistivity* property class.
     """
@@ -776,20 +628,6 @@ class r(PropertyInterface):
         return [T_m0, 1273.0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Electrical resistivity unit
-        """
-        return "[Ohm*m]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Electrical resistivity long name
-        """
-        return "electrical resistivity"
-
-    @property
     def description(self) -> str:
         """
         str : Electrical resistivity description
@@ -797,7 +635,7 @@ class r(PropertyInterface):
         return f"Liquid lead {self.long_name}"
 
 
-class k(PropertyInterface):
+class k(ThermalConductivityInterface):
     """
     Liquid lead *thermal conductivity* property class.
     """
@@ -833,20 +671,6 @@ class k(PropertyInterface):
         conductivity correlation function
         """
         return [T_m0, 1300.0]
-
-    @property
-    def units(self) -> str:
-        """
-        str : Thermal conductivity unit
-        """
-        return "[W/(m*K)]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Thermal conductivity long name
-        """
-        return "thermal conductivity"
 
     @property
     def description(self) -> str:
